src/push_products/requirements.txt
dataclasses==0.6
functions-framework==3.*
google-cloud-logging==3.11.*
google-cloud-bigquery==3.29.*
google-cloud-logging==3.11.*
google-cloud-tasks==2.19.*
grpcio==1.60.1
